--- a/mllib/src/main/scala/org/apache/spark/mllib/clustering/LDAOptimizer.scala
+++ b/mllib/src/main/scala/org/apache/spark/mllib/clustering/LDAOptimizer.scala
@@ -501,7 +501,7 @@ final class OnlineLDAOptimizer extends LDAOptimizer {
     logphatOption.foreach(updateAlpha(_, batchSize))
 
     expElogbetaBc.destroy(false)
-    stats.unpersist()
+
     this
   }