Hi, 

# What changes were proposed in this pull request?

as it was proposed by @jkbradley , ```gammat``` are not collected to the driver anymore. 

# How was this patch tested?
existing test suite.