@WeichenXu123 2017-09-13
ping @akopich This is an very useful improvement. Can you update the code while you're at it ?