@akopich 2017-09-20
@jkbradley, thank you for your comments! Please, check out the commit adding the necessary docs.

Regarding tests: I believe, `OnlineLDAOptimizer alpha hyperparameter optimization` from `mllib/clustering/LDASuite.scala` covers the piece of code being rewritten. Or should there be tests of a different kind?