@akopich 2017-09-23
@jkbradley, thanks for the comments. Who is supposed to create the followup jira?