@hhbyyh 2017-10-05
Yes, I think local test is enough for both correctness and performance. 

For consistency with old LDA, just some manual local test would be sufficient. You may well just use the LDA example and switch the Spark jar files between Spark 2.2 and your branch. And I think the case with empty document worth special attention.

The same for the performance test. You may just post the result after your local test. I'm OK as long as there's no noticeable regression.