@akopich 2017-10-05
Thank you, @hhbyyh.

I have augmented the example a bit: explicitly set random seed a nd chosen online optimizer:

`val lda = new LDA().setK(10).setMaxIter(10).setOptimizer("online").setSeed(13)`

But for some reason if I run it twice, the results are not the same. Is that expected? branch-2.2 was used.