@akopich 2017-10-05
I have conducted some performance testing with random data. 

The new implementation turns out to be notably faster. 

```
OLD with hyper-parameter optimization  : 237 sec
OLD w/o  hyper-parameter optimization  : 226 sec
NEW with hyper-parameter optimization : 178 sec
NEW w/o  hyper-parameter optimization : 171 sec
```