@WeichenXu123 2017-10-11
@akopich  LGTM. and do you have time to create a PR to resolve random seed not working issue mentioned by @hhbyyh ?  Thanks!