@akopich 2017-10-11
@WeichenXu123, yes sure. But can this wait until this PR is merged?