@jkbradley 2017-10-17
LGTM
Sorry for the delay!
I'll merge it after re-running tests