@akopich 2017-10-18
ping @jkbradley. Anyway, tests are passed now.