@jkbradley 2017-10-18
@akopich I'm afraid pings on Git don't work for me; I just have too many to keep up with.  Again, sorry for the delays; I have very limited bandwidth nowadays.